scanner fix

+++ common/scanner.h
@@ -3,8 +3,8 @@
 
 #include "p600.h"
 
-int scanner_keyState(uint8_t key);
-int scanner_buttonState(p600Button_t button);
+int8_t scanner_keyState(uint8_t key);
+int8_t scanner_buttonState(p600Button_t button);
 
 void scanner_init(void);
 void scanner_update(void);
